# packages/contracts/openapi/midas.openapi.yaml
openapi: 3.1.0
info:
  title: Midas Marketplace API
  version: 0.1.0
  description: >-
    Contrato executável dos cortes de identidade, SellerAccount e finanças:
    capabilities fail-closed, confirmação de pagamento por webhook, resolução
    administrativa auditável, saldo do seller e payout manual com segregação.
servers:
  - url: http://localhost:3001
paths:
...
  /v1/finance/provider-capability:
    get:
      operationId: getPaymentProviderCapability
      tags: [Finance]
      description: >-
        Informa a capability efetiva do adapter de pagamento. A ausência de
        provider selecionado, contrato, credencial ou adapter nunca é tratada
        como disponibilidade.
      parameters:
        - name: providerCode
          in: query
          required: false
          schema: { type: string, minLength: 1, maxLength: 120 }
      responses:
        '200':
          description: Capability de pagamento avaliada de forma fail-closed.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/PaymentProviderCapability' }
  /v1/provider-webhooks/{providerCode}:
    post:
      operationId: receivePaymentProviderWebhook
      tags: [ProviderWebhooks]
      security: []
      description: >-
        Recebe bytes originais do webhook e delega validação de assinatura e
        parsing ao adapter registrado. Não confirma pagamento sem provider
        disponível e evento verificável.
      parameters:
        - name: providerCode
          in: path
          required: true
          schema: { type: string, minLength: 1, maxLength: 120 }
      requestBody:
        required: true
        content:
          '*/*':
            schema: {}
      responses:
        '202':
          description: Evento autenticado recebido, inclusive replay idempotente.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ProviderWebhookAck' }
        '409': { $ref: '#/components/responses/Problem' }
        '422': { $ref: '#/components/responses/Problem' }
        '503': { $ref: '#/components/responses/Problem' }
  /v1/finance/payout-capability:
    get:
      operationId: getPayoutExecutionCapability
      tags: [Finance]
      description: >-
        Avalia a execução de payout manual externo pelo país, moeda e modo,
        sem presumir contrato ou credencial inexistente.
      parameters:
        - name: countryCode
          in: query
          required: true
          schema: { type: string, pattern: '^[A-Z]{2}$' }
        - name: currency
          in: query
          required: true
          schema: { $ref: '#/components/schemas/Currency' }
        - name: mode
          in: query
          required: false
          schema: { const: EXTERNAL_MANUAL, default: EXTERNAL_MANUAL }
      responses:
        '200':
          description: Capability de payout avaliada de forma fail-closed.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/PayoutExecutionCapability' }
  /v1/me/payments:
    get:
      operationId: listCurrentBuyerPayments
      tags: [Finance]
      security: [{ sessionCookie: [] }]
      responses:
        '200':
          description: Pagamentos vinculados ao comprador autenticado.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/FinancePaymentList' }
        '401': { $ref: '#/components/responses/Problem' }
  /v1/seller-accounts/{sellerAccountId}/finance/balance:
    get:
      operationId: getSellerFinancialBalance
      tags: [Finance]
      security: [{ sessionCookie: [] }]
      parameters:
        - $ref: '#/components/parameters/SellerAccountId'
        - name: currency
          in: query
          required: false
          schema: { $ref: '#/components/schemas/Currency' }
      responses:
        '200':
          description: Saldo financeiro canônico no escopo autorizado do seller.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/SellerBalance' }
        '401': { $ref: '#/components/responses/Problem' }
        '404': { $ref: '#/components/responses/Problem' }
  /v1/seller-accounts/{sellerAccountId}/payouts:
    parameters:
      - $ref: '#/components/parameters/SellerAccountId'
    get:
      operationId: listSellerPayouts
      tags: [Finance]
      security: [{ sessionCookie: [] }]
      responses:
        '200':
          description: Solicitações de payout visíveis ao seller autorizado.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/PayoutRequestList' }
        '401': { $ref: '#/components/responses/Problem' }
        '404': { $ref: '#/components/responses/Problem' }
    post:
      operationId: requestSellerPayout
      tags: [Finance]
      security: [{ sessionCookie: [] }]
      parameters:
        - $ref: '#/components/parameters/CsrfHeader'
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CreatePayoutRequest' }
      responses:
        '201':
          description: Solicitação de payout criada sem executar transferência.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/PayoutRequest' }
        '401': { $ref: '#/components/responses/Problem' }
        '403': { $ref: '#/components/responses/Problem' }
        '404': { $ref: '#/components/responses/Problem' }
        '409': { $ref: '#/components/responses/Problem' }
        '422': { $ref: '#/components/responses/Problem' }
  /v1/admin/payment-resolution-cases:
    get:
      operationId: listPaymentResolutionCases
      tags: [FinanceAdmin]
      security: [{ sessionCookie: [] }]
      parameters:
        - $ref: '#/components/parameters/FinanceStatus'
      responses:
        '200':
          description: Fila administrativa de casos de resolução de pagamento.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/PaymentResolutionCaseList' }
        '401': { $ref: '#/components/responses/Problem' }
        '403': { $ref: '#/components/responses/Problem' }
  /v1/admin/payments/{paymentId}/resolution-cases:
    post:
      operationId: createPaymentResolutionCase
      tags: [FinanceAdmin]
      security: [{ sessionCookie: [] }]
      parameters:
        - name: paymentId
          in: path
          required: true
          schema: { $ref: '#/components/schemas/FinanceUuid' }
        - $ref: '#/components/parameters/CsrfHeader'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CreatePaymentResolutionCase' }
      responses:
        '201':
          description: Caso auditável aberto sem editar o pagamento diretamente.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/PaymentResolutionCase' }
        '401': { $ref: '#/components/responses/Problem' }
        '403': { $ref: '#/components/responses/Problem' }
        '404': { $ref: '#/components/responses/Problem' }
        '409': { $ref: '#/components/responses/Problem' }
        '422': { $ref: '#/components/responses/Problem' }
  /v1/admin/payment-resolution-cases/{resolutionCaseId}/decisions:
    post:
      operationId: decidePaymentResolutionCase
      tags: [FinanceAdmin]
      security: [{ sessionCookie: [] }]
      parameters:
        - name: resolutionCaseId
          in: path
          required: true
          schema: { $ref: '#/components/schemas/FinanceUuid' }
        - $ref: '#/components/parameters/CsrfHeader'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/DecidePaymentResolutionCase' }
      responses:
        '200':
          description: Decisão registrada; aprovação depende de lookup provider SETTLED.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/PaymentResolutionCase' }
        '401': { $ref: '#/components/responses/Problem' }
        '403': { $ref: '#/components/responses/Problem' }
        '404': { $ref: '#/components/responses/Problem' }
        '409': { $ref: '#/components/responses/Problem' }
        '422': { $ref: '#/components/responses/Problem' }
  /v1/admin/payouts:
    get:
      operationId: listAdminPayouts
      tags: [FinanceAdmin]
      security: [{ sessionCookie: [] }]
      parameters:
        - $ref: '#/components/parameters/FinanceStatus'
      responses:
        '200':
          description: Fila administrativa de payouts pendentes ou concluídos.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/PayoutRequestList' }
        '401': { $ref: '#/components/responses/Problem' }
        '403': { $ref: '#/components/responses/Problem' }
  /v1/admin/payouts/{payoutRequestId}/claim:
    post:
      operationId: claimAdminPayout
      tags: [FinanceAdmin]
      security: [{ sessionCookie: [] }]
      parameters:
        - $ref: '#/components/parameters/PayoutRequestId'
        - $ref: '#/components/parameters/CsrfHeader'
      responses:
        '200':
          description: Payout movido para revisão pelo operador autorizado.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/PayoutRequest' }
        '401': { $ref: '#/components/responses/Problem' }
        '403': { $ref: '#/components/responses/Problem' }
        '404': { $ref: '#/components/responses/Problem' }
        '409': { $ref: '#/components/responses/Problem' }
  /v1/admin/payouts/{payoutRequestId}/approve:
    post:
      operationId: approveAdminPayout
      tags: [FinanceAdmin]
      security: [{ sessionCookie: [] }]
      parameters:
        - $ref: '#/components/parameters/PayoutRequestId'
        - $ref: '#/components/parameters/CsrfHeader'
      responses:
        '200':
          description: Payout aprovado para execução posterior.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/PayoutRequest' }
        '401': { $ref: '#/components/responses/Problem' }
        '403': { $ref: '#/components/responses/Problem' }
        '404': { $ref: '#/components/responses/Problem' }
        '409': { $ref: '#/components/responses/Problem' }
  /v1/admin/payouts/{payoutRequestId}/executions:
    post:
      operationId: registerExternalAdminPayoutExecution
      tags: [FinanceAdmin]
      security: [{ sessionCookie: [] }]
      parameters:
        - $ref: '#/components/parameters/PayoutRequestId'
        - $ref: '#/components/parameters/CsrfHeader'
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/RegisterExternalPayoutExecution' }
      responses:
        '200':
          description: Referência e evidência da execução externa registradas.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/PayoutRequest' }
        '401': { $ref: '#/components/responses/Problem' }
        '403': { $ref: '#/components/responses/Problem' }
        '404': { $ref: '#/components/responses/Problem' }
        '409': { $ref: '#/components/responses/Problem' }
        '422': { $ref: '#/components/responses/Problem' }
  /v1/admin/payouts/{payoutRequestId}/confirmations:
    post:
      operationId: confirmExternalAdminPayout
      tags: [FinanceAdmin]
      security: [{ sessionCookie: [] }]
      parameters:
        - $ref: '#/components/parameters/PayoutRequestId'
        - $ref: '#/components/parameters/CsrfHeader'
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/ConfirmExternalPayout' }
      responses:
        '200':
          description: Payout confirmado como pago por ator segregado e capability disponível.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/PayoutRequest' }
        '401': { $ref: '#/components/responses/Problem' }
        '403': { $ref: '#/components/responses/Problem' }
        '404': { $ref: '#/components/responses/Problem' }
        '409': { $ref: '#/components/responses/Problem' }
        '422': { $ref: '#/components/responses/Problem' }
        '503': { $ref: '#/components/responses/Problem' }
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: midas_session
  parameters:
    CsrfHeader:
      name: X-Midas-CSRF
      in: header
      required: true
      schema: { const: '1' }
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: true
      description: Chave estável por operação lógica; mínimo de 16 caracteres.
      schema: { type: string, minLength: 16 }
    SellerAccountId:
      name: sellerAccountId
      in: path
      required: true
      schema: { type: string, pattern: '^sac_[0-9a-f-]{36}$' }
...
    FinanceStatus:
      name: status
      in: query
      required: false
      schema: { type: string, minLength: 1, maxLength: 40 }
  responses:
    Problem:
      description: Erro tipado.
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
  schemas:
...
    Currency:
      type: string
      pattern: '^[A-Z]{3}$'
...
    PaymentProviderCapability:
      type: object
      additionalProperties: false
      required: [providerCode, status, reasonCode]
      properties:
        providerCode: { type: [string, 'null'], minLength: 1, maxLength: 120 }
        status: { enum: [AVAILABLE, CONTRACT_REQUIRED, UNSUPPORTED] }
        reasonCode:
          enum:
            - AVAILABLE
            - PROVIDER_CONTRACT_NOT_SELECTED
            - PROVIDER_CREDENTIALS_REQUIRED
            - PROVIDER_ADAPTER_NOT_INSTALLED
    ProviderWebhookAck:
      type: object
      additionalProperties: false
      required: [duplicate, processingResult, paymentId]
      properties:
        duplicate: { type: boolean }
        processingResult:
          enum:
            - SETTLED
            - PAYMENT_QUARANTINED
            - IGNORED_PENDING
            - IGNORED_FAILED
            - PAYMENT_NOT_FOUND
        paymentId:
          oneOf:
            - $ref: '#/components/schemas/FinanceUuid'
            - type: 'null'
...
    FinancePaymentList:
      type: object
      additionalProperties: false
      required: [data, asOf]
      properties:
        data:
          type: array
          items: { $ref: '#/components/schemas/FinancePayment' }
        asOf: { type: string, format: date-time }
    SellerBalance:
      type: object
      additionalProperties: false
      required:
        - sellerAccountId
        - currency
        - heldAmountMinor
        - availableAmountMinor
        - reservedAmountMinor
        - asOf
      properties:
        sellerAccountId: { $ref: '#/components/schemas/FinanceSellerAccountId' }
        currency: { $ref: '#/components/schemas/Currency' }
        heldAmountMinor: { type: string, pattern: '^[0-9]+$' }
        availableAmountMinor: { type: string, pattern: '^[0-9]+$' }
        reservedAmountMinor: { type: string, pattern: '^[0-9]+$' }
        asOf: { type: string, format: date-time }
    PayoutExecutionCapability:
      type: object
      additionalProperties: false
      required: [countryCode, currency, mode, status, reasonCode]
      properties:
        countryCode: { type: string, pattern: '^[A-Z]{2}$' }
        currency: { $ref: '#/components/schemas/Currency' }
        mode: { const: EXTERNAL_MANUAL }
        status: { enum: [AVAILABLE, CONTRACT_REQUIRED, UNSUPPORTED] }
        reasonCode:
          enum:
            - AVAILABLE
            - PAYOUT_CONTRACT_NOT_SELECTED
            - PAYOUT_CREDENTIALS_REQUIRED
            - PAYOUT_MODE_UNSUPPORTED
    PayoutRequest:
      type: object
      additionalProperties: false
      required:
        - payoutRequestId
        - sellerAccountId
        - requestedByUserId
        - amountMinor
        - currency
        - destinationCountry
        - payoutStatus
        - claimedByUserId
        - approvedByUserId
        - completedByUserId
        - claimedAt
        - approvedAt
        - completedAt
        - createdAt
        - version
      properties:
        payoutRequestId: { $ref: '#/components/schemas/FinanceUuid' }
        sellerAccountId: { $ref: '#/components/schemas/FinanceSellerAccountId' }
        requestedByUserId: { $ref: '#/components/schemas/FinanceUserId' }
        amountMinor: { $ref: '#/components/schemas/AmountMinor' }
        currency: { $ref: '#/components/schemas/Currency' }
        destinationCountry: { type: string, pattern: '^[A-Z]{2}$' }
        payoutStatus:
          enum:
            - REQUESTED
            - UNDER_REVIEW
            - INFORMATION_REQUIRED
            - APPROVED
            - REJECTED
            - EXECUTING
            - CONFIRMATION_PENDING
            - PAID
            - FAILED
            - RETURNED
            - CANCELED
        claimedByUserId:
          oneOf:
            - $ref: '#/components/schemas/FinanceUserId'
            - type: 'null'
        approvedByUserId:
          oneOf:
            - $ref: '#/components/schemas/FinanceUserId'
            - type: 'null'
        completedByUserId:
          oneOf:
            - $ref: '#/components/schemas/FinanceUserId'
            - type: 'null'
        claimedAt: { type: [string, 'null'], format: date-time }
        approvedAt: { type: [string, 'null'], format: date-time }
        completedAt: { type: [string, 'null'], format: date-time }
        createdAt: { type: string, format: date-time }
        version: { type: integer, minimum: 1 }
    PayoutRequestList:
      type: object
      additionalProperties: false
      required: [data, asOf]
      properties:
        data:
          type: array
          items: { $ref: '#/components/schemas/PayoutRequest' }
        asOf: { type: string, format: date-time }
    CreatePayoutRequest:
      type: object
      additionalProperties: false
      required: [amountMinor, currency, destinationCountry]
      properties:
        amountMinor: { $ref: '#/components/schemas/AmountMinor' }
        currency: { $ref: '#/components/schemas/Currency' }
        destinationCountry: { type: string, pattern: '^[A-Z]{2}$' }
...
    PaymentResolutionCaseList:
      type: object
      additionalProperties: false
      required: [data, asOf]
      properties:
        data:
          type: array
          items: { $ref: '#/components/schemas/PaymentResolutionCase' }
        asOf: { type: string, format: date-time }
...